A more sophisticated deploy.sh
It now controls what exactly is synced very carefully, using
rsync --delete --delete-excluded. A temporary directory is used
for syncing to which is then copied and used to replace the original
target folder using (nearly atomic) moves.

Data which must persist on the server, such as .env and uploads, are kept
in a separate directory and symlinked in following a deployment update.

Deploying to a staging test dir is now supported too.