Bring back shelljs for executing shell commands
When I removed it, I wondered why shelljs was creating files for stderr, stdout and parameters... I now figured out that it's the most robust way to handle lots of data without hitting the maxBuffer limit.

`netlifyctl` in particular is pretty noisy and easily reaches the maxBuffer limit, which also means all deploys will fail with the existing code.