Add foreign keys for chart_revisions & chart_slug_redirects
Ideally, chart_revisions should also have a chartId foreign key, but then charts that get deleted must have all revisions deleted. It's better to not enforce this constraint.

In the future, we should really just *mark* charts as deleted, not delete the chart rows. Same applies for the other tables.