Store `isExplorable` in database
Note that it is stored both in a separate database column and in the chart config object.

I think it's preferred that it is a column instead of buried in a JSON object, and we shouldn't have it duplicated in both places. But it's fiddly to ensure that all endpoints that return a chart inject an isExplorable property. So I opted to have it duplicated.