Force some charts to be explorable
The live server does not yet have the checkbox to make a chart explorable, it is hidden behind a feature flag. So whenever we overwrite local/staging with the live database, we currently lose the explorer charts.

In order to avoid remarking the charts, I've added some chart IDs to be forced to be explorable, even if their config does not have the flag set.