fix: let footnote hover area overflow visibly
Horizontal scroll bar would appear when footnote at the end of a line and whose hover area (:after) would overflow. overflow-x: visible would actually be interpreted as auto, as per following note from MDN: "Note: Setting one axis to visible (the default) while setting the other to a different value results in visible behaving as auto."
Tables are not in <p> (anymore?) so the historical comment is no longer relevant.