Fix typechecking on first() and last()
Our utils didn't acknowledge that these can return undefined when the
array is empty. One downstream consequence of this was a bunch of
console warnings about trying to format an invalid year of undefined.

The one thing I left in a funky state is Timeline, because it makes too
many assumptions. Instead, I tried not to render it if the years array
is empty, and I put a warning on the constructor if this is ever missed.
So far, no warnings in tests or on the console in dev.