Consistent header dropdown label
The site header is defined in two places – one has interactive behaviour, one is used in the baker. This is kind of stupid obviously, and I don't remember the details of why we did it this way, and we should fix it at some point, but just keeping them in sync for now.