Netlify: add explicit ! to (implicit) forced redirects
On April 7th, 2020 Netlify will change the way they handle redirects. Currently, the default behaviour is to always assume redirects are forced – a path will be redirected even if the file exists.

This behaviour will change on April 7th, and we will need to explicitly mark our redirects with a trailing `!` in order to preserve the current behaviour.

More info: https://community.netlify.com/t/changed-behavior-in-redirects/10084