Bake chart pages from existing pre-fetched config
(Probably) fixes the issue where charts don't refresh their data on republication.

The config for all charts is fetched and cached by the baker before chartPages are baked. But the baking of chart pages used to query the database to render the page, which meant that a chart page could receive the new config, but the change in version goes undetected because it happened in the period between when the prefetch happened, and the subsequent query to bake the individual chart page.