Change ControlsOverlay id
This id is only used for keeping track which portal belongs to which component, so that it can be unloaded when the component unloads. Has a very similar role to React's `key` prop.

This change makes no difference since there is no other "add-country" component rendered while this is rendered, but it's good to keep them distinct.

There is probably a better way to implement the overlay association than hand-crafted ids (with portals maybe?) but I didn't manage to figure it out in the past...