Pass unbounded time to Map timeline
Fixes a bug where opening a Map in the Admin would turn 'Infinity' into the latest available date. This is because `map.data.targetYear` is of the Time type, so it cannot be unbounded – any unbounded collapses to a specific time. `map.targetYear` on the other hand is a TimeBound type – the one we need here.