Fix & generalize Explorer redirects
Now redirects are no longer Grapher-specific, we can add redirects from any path on the site (this is how "/coronavirus-data-explorer" is implemented).

The redirects are all in a CSV, and one of the columns is `migrationId`. Currently the only `migrationId` we have is `legacyToGridCovidExplorer` (see variable `explorerUrlMigrationsById`).

In the future, we can add others! This could be the way we implement backwards-compatible URLs while we change the Explorer grid programs.