fix: use serverSettings in server-side code
otherwise, staging sites were showing links like `localhost:3030`
or `owid.cloud` instead of their staging-server equivalents.
this is because `clientSettings` are not filled in for server-side code.
I'm still not happy with all of this and it's still too easy to mess
this up by accident, so maybe we should look at doing this in a better
way once again.