fix(grapher): Line charts with projections were broken
This fixes the issue described in https://www.notion.so/owid/Broken-chart-ea6e5f3f9dcf49d4859b74dad585d6ca). The problem is that autoDetectSeriesStrategy checks if there is more than one variable involved and if so it sets the strategy to column. When there is a projection we have two variables but logically this is shown as one and entity is still the right strategy.