chore(grapher): Add return type annotations
This commit re-enables several TS lint rules project wide. The two warnings related to the return type annotations create the most warnings. These are disabled on the project level for now and re-enabled on specific folders where the annotations have been added. About 70% of the grapher is annotated correctly now as are a few of the shared folders. The admin client and server are still missing.

This commit should not change any functionality and only add more type safety. A few select places of the code where changed to make type annotations easier/more understandable (e.g. use .? instead of && in some places)

This commit has been squashed and represents several individual PRs and feedback and help contributed by Daniel G and Marcel.