fix: make country selection top-to-bottom, left-to-right
Switch from flex box for entity selector to using ul columns with a
media query.

This is because countries were spilling left-to-right before top-to-bottom,
a highly unintuitive sort order for people to look through. A media
query approach was taken because flexbox columns would only solve this
if the container had fixed height.

https://www.notion.so/owid/List-countries-vertically-instead-of-by-rows-in-grapher-country-selector-ad9fbd8b86c04e61bf58b68dd5e1bade