enhance: Shrink explorer title to reduce wasted space
In a previous design, the title could take more space because there were
more controls to the right of it. Now that there are fewer, there is
blank space left.

This change reduces the title so as to remove the need for blank space
in the panel to its right.

See: https://www.notion.so/owid/Condense-explorer-title-d5bbac463dd54470a60cac48c1c2dee2