fix(linechart labels): Extend line chart labeling logic
The fix for projections in line charts meant that sometimes the "- Projected" part of the label was dropped due to the flags influencing label generation. This commit fixes this for most cases. There remain a few charts that have slightly odd labels but I think this is an improvement over the previous state overall.