fix: hide timeline for Source or Download tabs
When you click on the Source or Download tabs of a chart, the timeline
currently stays rendered, despite logic that is meant to hide it.

This is because the check is done on `this.tab`, but actually Source and
Download are overlays on the tab; when you click on them, `this.tab`
stays at wherever it was last.

This fix runs the same check, but on `this.currentTab` instead, which
picks up any overlay (such as Source or Download) if they are available.
This has the effect of correctly hiding the timeline when the user
clicks on those tabs.

See: https://www.notion.so/owid/Hide-timeline-on-Sources-and-Download-tabs-8766e5c1c13f431b978b8d579b92ff65