refactor: drop `yAxis` and `xAxis` from `ChartManager`
There used to be two ways to pass axis config: with `yAxisConfig` which is a plain object, and with `yAxis` which is a class instance. The DiscreteBar had a bug because it didn't consider `yAxisConfig` in its implementation. Now with only one way of passing axis config down to charts, there is less chance we make a mistake.