chore: add a short tmux cheatsheet to the welcome banner
The way tmux is being used is slightly non-standard, where some panes
will not close if the process crashes. Instead, they sit open and you
have the option to restart them with <C-b>, R. In this way, you don't
lose an important pane just because it crashed.

For this reason, some extra tmux bindings were added to the Makefile,
and a cheatsheet is added to the welcome banner.