chore(actions): don't use secret GH token for external run of svg-compare action
The secret token is not available when the action is run in an external
PR.
This way, we're only using the secret token when running internally.
When running from a fork, the checkout of owid/owid-grapher-svgs
succeeds even with the public token, and the SVG tester will run fine
_if there are no SVG differences_. If there are some, the action will
fail without pushing the changes up.