chore: Pin the id of the dev user to 1
The dev user currently gets inserted and auto-assigned an integer `id`.
But other tools like the `etl` and `importers` that need to use the
grapher DB need to specify a user id, and we would like them to use this
dev user for their example config.

Currently id 1 is unused in production (id 2 is Max).