enhance(explorer): improve dynamic page titles exposed to Google
We have an open experiment where we can opt in explorers to be expanded
in our sitemap (#1214). The purpose of this expansion is to allow people
to arrive, through Google Search, at specific views of our data
represented by specific explorer configurations.

A current problem is that our page titles are all too similar for
explorer views, so Google omits them from search as being not helpful to
the user.

This change modifies the page title each time the explorer view changes
from the default setting, for opted-in explorers. The page title chosen
is the grapher display title suffixed with "- Our World In Data".