fix: Add updatedAtInWordpress field to posts table
Because the posts table mirrors wordpress posts into our database,
we use it to understand when the wordpress post was last modified.
The recent unification of automatic timestamps interferes with this
now - when running the archieml update for example that touches
every post, the updated_at field is now automatically changed.
This commit adds a new field so that we can separate the "wordpress
changed" meaning from "the table was touched"