fix: use `axis.invert` for mouse tooltips
I noticed a rare issue with #1420: for charts with super-long domains,
the use of `minBy` would make it impossible to see a tooltip for the
_last_ year.
This was happening on http://ourworldindata.org/explorers/climate-change?Metric=CO%E2%82%82+concentrations&Long-run+series%3F=true,for example,
which probably is our longest time series.
Using `axis.invert()` fixes this, and is probably also faster overall.