feat: add an unlink button - this keeps the gdoc but severs the link to posts/gdocs
This was requested by Joe.  It happens that authors
are working on converting an entry that is not
supposed to be converted as one entity. In
those cases it is useful if, after creating a gdoc
and copying some fragments, the link to the
gdoc can be cut again so as not to confuse
yourself or others in the future.