chore: use a sensible default for the image bucket path
At the moment, when you run `make up.full` for the first time, it copies
an template `.env` file for you that should work out of the box, except
one setting must be set manually, the image bucket path.

This change sets it to `owid-image-upload/dev-<user>`, where `user` is
the unix username of the current user.