fix: do not push explorer changes in dev
Currently, when you modify explorers using the admin UI, it immediately
writes and pushes these changes to production, regardless whether you
are in a dev or staging environment.

This change fixes this for dev environments, by modifying the
GitCmsServer to avoid doing Git operations (commit, push) if in a
development environment.

Ideally we would do the same for staging. This would need us to be able
to easily differentiate between prod and staging, which we can't easily
do yet.