:honeybee: set the dev tools to use a profile for s3 credentials
Previously, our dev tool used the default s3 credentials when syncing to
DigitalOcean Spaces. This is a little problematic, since we are now
using multiple object stores.

This change suggests setting up a profile for OWID and explicitly using
that. In this way, it's compatible with using multiple object stores,
and in fact having multiple sets of S3 credentials on your machine.