Updated images and text to match what I see on a fresh install
These things sometimes get out of date, so I thought I'd update - we've got 5 TMUX panels now and one for the DB, rather than having it in docker. 

I updated the image, and provided some text because data just in images is bad, plus, upgraded the headiong of "trouble shooting" so it doesn't look like a sub of the section before, really, troubleshooting is it's own special thing.

Updated a bit of text - do check this, I'm new.