Use Playfair Semibold in static grapher exports (#3225)
The png-generation code in StaticRasterizer needs to preload any fonts it uses and had previously been expecting to use Playfair Medium for chart titles. The updated design involved changing the title weight to Semibold but neglected to update the rasterizer. This PR brings the rasterizer and static chart code in sync.

There are a number of pre-generated static exports that need to be re-rendered to use the updated typography. Since the baking process only generates a new export if the config has changed since the previous bake, we need to force a full re-render manually.

One possibility would be:
1. hard-coding the [`chartIsSameVersion` function in GrapherBaker.ts](https://github.com/owid/owid-grapher/blob/fd511a1b6c254f12a08ffa2c20906b70f981220a/baker/GrapherBaker.tsx#L384-L396) to return `false`
2. triggering a bake
3. restoring the normal lazy behavior of `chartIsSameVersion`

------
closes #3210
incorporates #2970