:hammer: Add "make bench.search" to evaluate search performance
It fetches a dataset of synthetic queries and evaluates the extent to
which we surface good articles for the given queries.

The scoring algorithm chosen for articles is `precision@4`, meaning the
the proportion of the first four results that are relevant, averaged
over a ton of queries.

This is chosen since at most four articles are
presented un-collapsed, and the value of getting those four right is
much much higher than getting any right further down in the ranking.

It does not yet score chart or explorer search.