✨ (entity selector) remove slide-in settings drawer (#3353)
[Cycle 2024.2: Entity selector](https://github.com/owid/owid-grapher/issues/3349) (prep work)

Removes the slide-in settings drawer and always renders settings into a popup instead.

We will be using the slide-in drawer for the entity selector instead (at certain screen sizes). I removed the slide-in drawer code in this PR, but I will be re-adding it for the entity selector in subsequent PRs.