🔨 (data page) use full-width grid (#3624)
The data page layout is padded on both sides. I removed the padding in favour of a full-width grid layout (similar to how it's done on topic pages). As a result, the margins on the side are 16px on smaller devices (not 24px), which is more consistent with the designs and also allows me to make mobile charts full-width in a second step :)

| Topic page  | Data page  |
| ------- | ------ |
|  <img width="507" alt="Screenshot 2024-05-21 at 14 51 49" src="https://github.com/owid/owid-grapher/assets/12461810/a6a1f320-e67f-4de8-a495-bcb4cadf9e41"> | <img width="507" alt="Screenshot 2024-05-21 at 14 51 30" src="https://github.com/owid/owid-grapher/assets/12461810/787b4459-388a-4b81-b1aa-870bdb8c78b6"> |