✨ (grapher) improve line legend algorithm (#3647)
Resolves https://github.com/owid/owid-grapher/issues/3351

### Current implementation

The current implementation uses two different placing algorithms:
  1. If all labels fit into the space, then labels are placed in a non-overlapping way using connector lines
  2. If the labels don't fit into the available space, then labels are simply placed at their y-value and overlapping labels are greyed out and rendered into the background

### Updated implementation

I got rid of the second placing algorithm. Instead, if not all labels fit into the available space, then we drop as many labels as necessary, and then the first placing algorithm is used to compute the positions of the remaining labels.

#### How labels are dropped

The default strategy, used by line charts, tries to pick labels in a balanced way so that they're roughly at equidistance from each other. Stacked area charts overwrite this behaviour and drop labels based on the area size.

The default strategy works like this: Given a set of placed labels and a set of candidates, for each candidate, we find the two closest already placed labels, one to each side, and calculate a score based on the available space between the two placed labels (the bigger, the better) and the candidate's distance to the midpoint (the smaller, the better). We then pick the candidate with the best score that fits into the available space.

### Future Work

Lines/Areas that don't get labelled should show a label on demand (previously, you could hover over the grayed-out background labels). Stacked area charts highlight the currently hovered area in the tooltip – we could do something similar for line charts. To be worked on...

| Before  | After  |
| ------- | ------ |
|  ![total-population-living-in-extreme-poverty-by-world-region-mobile (1)](https://github.com/owid/owid-grapher/assets/12461810/71f8a036-784d-4ae1-a146-096bdc47f9d4) | ![total-population-living-in-extreme-poverty-by-world-region-mobile](https://github.com/owid/owid-grapher/assets/12461810/0c7c76c5-5869-4871-a7de-87024ed96e48) |
| <img width="317" alt="Screenshot 2024-05-28 at 16 15 48" src="https://github.com/owid/owid-grapher/assets/12461810/325ab991-4cc3-4b41-a9c9-c94c69126765"> |<img width="317" alt="Screenshot 2024-05-28 at 16 36 15" src="https://github.com/owid/owid-grapher/assets/12461810/5f2cf38c-1419-47af-a4b8-085e39c718e6">|