chore(deps): add `colorjs.io`
this module is by Lea Verou, co-editor of the CSS Color 4 and 5 specs, and also has some advanced color manipulation features that may be useful to us in the future.
Include chromatic adaptation, which we might be able to use an alternative whitepoint? not sure yet if that helps in any way