🎉 (static charts) add output channel for social media / TAS-639 (#3993)
Adds a flag to download a static chart optimised for social media / Instagram. The option is accessible from Grapher's download menu and is only available for staff members who are logged in.

This is how the social media export is currently different from the ordinary export:
1. Background is beige (+ line outlines etc are also beige)
2. Background `<rect />` not needed in Figma is removed
3. Dash pattern of grid lines is set to `7 7` (Lucas prefers this)

To be honest, I'm not sure if adding a flag for this special behaviour is worth it. Alternatively, we could allow to customize Grapher's background colour. Maybe that's nicer? We could also do none of these things and that would be fine as well! Curious what you think, @danyx23 

SVG tester changes come from the previous PR

<img width="838" alt="Screenshot 2024-10-03 at 14 45 51" src="https://github.com/user-attachments/assets/0d0d97b2-5e03-47dd-9126-910b25d8b200">