✨ (scatter) don't animate as a time scatter (#4038)
Animates a scatter as a cloud of moving dots and keeps the axes fixed. The axes are fixed for the particular time range that is animated. 

At the animation's beginning and end, the view "jumps" to the new axis domains. The same happens when an animation is paused and the user starts dragging one of the sliders.

The animation in relative mode is unchanged. Getting the right axis domain doesn't seem trivial to me unless I'm missing something, so it's not worth doing at this point IMO.

In a follow-up PR, I'll also look into fixing the size domain.