🎉 (slope) add tolerance / TAS-719 (#4205)
Adds tolerance to slope charts.

If the original times of start and end value are too close together, then we don't show the slope. This is something that came out of discussion with Joe and Daniel, but now that I'm working on this, I don't think we need to be that strict? As long as the original start time < the original end time, we should be good? Curious what you think!

### Testing

Example charts:
- http://staging-site-slope-tolerance-viz/grapher/ratio-of-female-to-male-labor-force-participation-rates-slopes?time=2000..latest&country=AGO~OWID_WRL
- With forced relative mode (doesn't make much sense here, just for testing): http://staging-site-slope-tolerance-viz/grapher/ratio-of-female-to-male-labor-force-participation-rates-slopes?time=2000..latest&country=AGO~OWID_WRL&stackMode=relative