✨ Add gdoc_components_table that stores flat components from all docs
This PR adds a new table that contains an unfolded, flat list of all gdocs components in all gdocs.

For each gdoc, the tree in $.content.body is iterated, a copy of the content is made without any children and the children are recursed into. Span arrays are converted to plain text.

A new script, reconstructPostsGdocsComponents fills the db initially. When saving gdocs, the components for this gdoc are updated in the new posts_gdocs_components table.