🔨 (line legend) refactor label dropping algorithm (#4310)
Refactors the code that is responsible for dropping labels if there is not enough space.

The SVG tester comes back with one additional difference. It's a slope charts where the 'Tanzania' and 'Mali' labels are swapped. Tanzania should be on top (its value is 1339) and 'Mali' should be below (its value is 1337), so the updated chart is correct.