✨ Move remaining computed props into grapherstate. This fixes a few bugs.
The main channel through which this led to undesired behavior is that we used to pass
the full Grapher as manager into components like the Footer; we now only pass the
GrapherState. But because the interfaces that outline managers (like that of the footer)
often use optional properties, the type system didn't flag that some of the computed
properties were missing now on GrapherState. This commit moves the remaining computed
state props onto GrapherState.