✨ Extracts most of Grapher's props into a new GrapherState class
This is a large refactor that splits the complex Grapher react component
that was also a complex MobX state class into two: a GrapherState mobx
class that is not a react component but deals with almost all the internal
state and calculations; and Grapher, which is a react component that
expects a GrapherState instance to be passed in via props (and nothing
else). The other big change is that data loading no longer
happens automatically from within Grapher - instead, the usage pattern
is to use a useRef hook to instantiate a GrapherState instance, deal
with data loading, and then set the inputTable on the GrapherState
instance in the owning component. A new component, FetchingGrapher
abstracts the data loading away.