Add site redirects for recent chart redirects
We take Grapher redirects into consideration only when a route returns
404 and Cloudflare caches assets up to a week in the CDN, so it's
possible to still get the old page after a redirect is created for it.

Authors are confused why they don't see chart redirects working after
they push changes. It also means different users can see different
charts when clicking on the same link, depending on whether the CDN data
center that handles their request had the old page cached. I tested this
case with Pablo A.

We take a similar approach as for deleting articles and add redirects in
the special `_redirects` file - that Cloudflare always takes into
consideration - for Grapher redirects that are less than a week old.