Fix logo position in them main nav on a small screen (#4955)
The logo wasn't centered on small screens.

The `column-gap` polyfill doesn't work correctly when the last child is
`display: hidden` which became the case at some point.

Went with this fix to be on the safe side, instead of switching to the
`column-gap` property, which wouldn't work correctly for all widths by
itself anyway.