Remove buildTsc script (#4989)
Since we switched to `tsx` to execute TS files directly we no longer
need to compile TS to JS explicitly before running any commands.

We now have the `typecheck` script, which currently does almost the same
as the old `buildTsc` (without the verbose flag), but it is used only
for checking the types and thus we don't need to run it in so many
places.