Fix spacing of logos in archive on small screen
There was a missing margin between the OWID logo and the affiliate logos
on archive site on a small screen introduced in 0324820c2.

We can simplify the code, but the change introduces smaller margin
between the OWID logo and the affiliate logos on live site on a single
size, which probably wasn't even intentional, since the margin is
smaller on both smaller and larger screen sizes.