Implement browser design for large topic page hit
This is initial prototype to get feedback on. It's not clear how will we
handle topic pages with long intros and linear topic pages, which don't
have any markup for intros.

Here I included a first attempt at a heuristic for that, which is a bit
crazy. Using normal excerpt for linear topic pages might be better.