Fix type of DbInsertPostGdocLink (#5416)
DB types should correspond to DB table schema definition, but for
`DbInsertPostGdocLink` the type of the `sourceId` column was also
number, despite it only being string | null in reality.

I also created a new helper `createDodLinkFromUrl` for DoD links that we
use in the admin API handler and the old DB migration and made the types
of the other helpers more specific.