Rename archive variables to better match types (#5395)
## Context

In the past, the type(s) got renamed without renaming all of their variables. Since we are going to start archiving articles, which will also use `ArchiveContext`, it makes even more sense now to stop using "chart" as part of variable names for `ArchiveContext` and clean up the naming.

This also fixes a few other instances that I found confusing.

This change should not affect the behavior in any way.

## Testing guidance

Check that everything still works as expected.