Fix entity selector drawer closing on dropdown click (#5547)
React Aria has an escape hatch for providing a portal container for
cases like this one. The nicer way of solving this would be by using a
React Aria component for the slide in drawer, which should then make the
dropdown behave correctly.

However, we have an indiosyncratic UI drawer that doesn't follow
established patterns (is not a modal) and it's not clear if it could be
implemented with something like RA Popover while preserving the current
behavior.