Switch SearchDropdown to searchable select
The main difference is in where the various parts are in the UI and how
they behave. Combobox is more of a text input + list of
suggestions (which you can see in full if you don’t type anything),
while searchable select is basically just a select with search field
included in the dropdown.

https://react-spectrum.adobe.com/react-aria/examples/searchable-select.html