Switch SearchDropdown to searchable select (#5560)
The main difference is in where the various parts are in the UI and how they behave. Combobox is more of a text input + list of suggestions (which you can see in full if you don’t type anything), while searchable select is basically just a select with search field included in the dropdown.

https://react-spectrum.adobe.com/react-aria/examples/searchable-select.html

## Context

[Slack](https://owid.slack.com/archives/C5BDCB2R3/p1760947370928079)

## Testing guidance

Please check that other than the inteded changes, the behavior is correct.