Serialize archive context only once
In SiteFooter we were serializing the archive context only on archived
pages, i.e. only when it was ArchiveMetaInformation. AFAICT this was
originally done for DoDs, which are handled outside of the normal
component tree and which don't need the archive context on live pages.

For live pages we also need the archive context to use an archive link
in citations and embeds. In this case the archive context is
ArchivedPageVersion. We were serializing it separately for different
kinds of pages, both for archived and live versions, which meant that
the same archive context was serialized twice for archived pages: once
in the footer and once in props of the specific page, which was
unnecessary.

Since we need to put it in a separate global prop on window for DoDs, we
can rely on that prop also for the other use cases and remove the
additional serialization.