Add a force option to archiveChangedPages
We could already force archiving of pages by specifying their id/slug
using the other existing options, but we now also need a way to force
archiving of all pages or all types of pages when combined with the
various options to select page types.

A follow-up to #5715 so we can re-archive all articles.